requirements.txt
boto3>=1.34.0
pandas>=2.0.0
numpy>=1.24.0
snowflake-connector-python>=3.6.0
apache-airflow>=2.8.0
